+++ NEWS.md
@@ -17,6 +17,7 @@
 * `overlap()`: the `which` param can now also be a `list` to compare groups of replicate groups (similar to `plotVenn()`)
 * Fixed: `plotSpectrum()` method for `formulas`/`compounds` didn't expand plot height for formula annotations with only one mass peak
 * Fixed: `annotatedPeakList()`/`plotSpectrum()` methods for `compounds` didn't label mass peaks with compounds algorithm if `formulas` were provided but no formula candidate was present.
+* `consensus()` method for `featureGroups`: new `verifyAnaInfo` flag to optionally skip if the analysis information are equal for all compared objects. This is mainly useful when the data is the same but in different formats.
 
 
 # patRoon 2.2